Move multi-device support into a MultiDeviceFitter subclass
The device layout is fixed at initialization and never changes during a
fit, so the single-device / multi-device split is a static choice best
expressed as a class: rabbit.sharding.MultiDeviceFitter subclasses
Fitter and owns all sharding -- shard construction, the per-device
tf.function assembly, the set_nobs propagation -- while the base Fitter
carries none of it. make_fitter() picks the class from --nDevices, so
call sites stay generic.

The base-class diff reduces to making the deepcopy strip-list an
extensible class attribute: since __deepcopy__ already re-instantiates
via type(self) and ends in _make_tf_functions(), and init_fit_parms
already ends in _make_tf_functions(), the subclass rebuilds its shards
by overriding that one method -- no hooks, no if-shards branches, and
toys preserve the subclass automatically. The likelihood mathematics
remains shared by construction: shard evaluators borrow the base
class's own unbound methods, so this class is orchestration only.

Behavior is unchanged: the full sharded and single-device test suites
pass identically (101 tests).

(cherry picked from commit 30bbfa3cc4241beb9a1d2cf75aff97d40d8ba382)
